+++ templates/workspace_services/guacamole/user_resources/guacamole-azure-linuxvm/terraform/deploy.sh
@@ -1,0 +1,18 @@
+#!/bin/bash
+
+set -o errexit
+set -o pipefail
+set -o nounset
+# Uncomment this line to see each command for debugging (careful: this will show secrets!)
+# set -o xtrace
+
+export TF_LOG=""
+# These variables are loaded in for us
+# shellcheck disable=SC2154
+terraform init -input=false -backend=true -reconfigure -upgrade \
+    -backend-config="resource_group_name=$TF_VAR_mgmt_resource_group_name" \
+    -backend-config="storage_account_name=$TF_VAR_mgmt_storage_account_name" \
+    -backend-config="container_name=$TF_VAR_terraform_state_container_name" \
+    -backend-config="key=${TF_VAR_tre_id}${TF_VAR_workspace_id}${TF_VAR_parent_service_id}guacamolelinuxvm"
+terraform plan
+terraform apply -auto-approve